Refactor cron.
