+++ go.mod
@@ -8,7 +8,7 @@ require (
 	github.com/pkg/errors v0.9.1
 	github.com/stretchr/testify v1.5.1
 	github.com/uber-go/tally v3.3.15+incompatible
-	go.temporal.io/temporal v0.10.14
+	go.temporal.io/temporal v0.10.15
 	go.temporal.io/temporal-proto v0.0.0-20200316214407-583dbd3e3b32
 	go.uber.org/zap v1.14.1
 	golang.org/x/sys v0.0.0-20200317113312-5766fd39f98d // indirect
